t/uni/variables.t:  Start to adapt to Perl 7
This file runs a loop over all ASCII characters (0x0..0xff).  In each
iteration of the loop, a signal handler pushes warnings onto an array.
At the end of each iteration the number and kind of unit tests run vary
depending on certain factors, one of which is the number of entries in
the array holding warnings for that iteration.

The problem is that code examining those warnings implicitly assumed
that each warning constituted a single line.  (This assumption was
probably made because we were interested in one specific warning.)
However, on the following ASCII characters ...

    0x0, 0x9 .. 0xd, 0x20

... *two-line* warnings are being emitted along the lines of:

    Number found where operator expected at (eval 55) line 1, near "$^@ = 4"
      (Missing operator before 4?)

For the numbers in question, the number of lines in the array holding
captured warnings is now non-zero, so we get a failure for each of these
tests.

As a step toward fixing this problem, let's start by excluding the
problematic characters from the while loop.  There are 7 such
characters, for each of which 6 unit tests were being run, so the plan
is reduced by 42.