t/uni/variables.t: Write separate 'for' loop for 7 characters emitting warnings
These are either 0x0 or the ::space:: characters.  Because (a) we
already expect these to emit syntax errors and (b) know what other regex
classes they belong to, we can jettison much of the infrastructure used
in the previous 'for' loop.  So, e.g., we don't need an 'if' block which
asks if we are expecting a syntax error or not; we simply proceed on the
basis that we are indeed expecting one.