dist/Exporter/t/warn.t: Scrap useless prototype; file now compiles
This file was gagging for two reasons:  The declaration of sub ok with
prototype ($;$) was causing all kinds of syntax errors.  And @EXPORT_OK
needs to be declared with 'our' once 'use strict' is in effect.  The
second of those is easily remedied.

As to the first, we could "cheat" and declare the subroutine this way:

    sub ok :prototype($;$) {
        ...
    }

But someone would be sure to point out that Exporter has a dual-life on
CPAN and this type of prototype definition would break perls earlier
than 5.18.

What if we asked ourselves, "What were we actually getting from the
prototype, any how?"  Very little.  All that the prototype was saying
was that the function takes two scalar arguments, the second of which is
optional.  The prototype doesn't even require the first argument to be
defined.  It is the function itself which examines whether the first
argument is defined or not -- and the function does not cause an
exception even in that case.

In short, the prototype definition is accomplishing nothing.  Junk it.