ext/SDBM_File/corrupt.t:  Correct likely typo
When run with warnings-on-by-default, one statement in this file was
causing two warnings to be emitted.

Upon reflection, this appears to be a problem in blead which would have
been detected if the file had said 'use warnings;'.  See:
https://github.com/Perl/perl5/issues/17970

Fixing the spelling means we don't have to suppress any warnings.