dumper.t:  Get file to compile
The purpose of this commit is to "clear away the underbrush" of warnings
and failures in dumper.t.  We honor strictures within the BEGIN block
and a few other places.  We overcome a few fatal errors. We suppress a
few warnings.  The file now runs to completion but has 150 failures --
which are real regressions from Perl 5.