parts/inc/inctools: Quiet uninitialized warnings in Perl 7
The prototype on sub dictionary_sort() was preventing compilation and
wasn't getting us anything.  In particular, that sorting function was/is
being called in many places where one or the other of the two arguments
was uninitialized -- but that's cool.  So let's jettison the prototype.

That enables dist/Devel-PPPort/t/ppphtest.t to compile and run, albeit
shakily.  Running that illustrated the need to further modify
dictionary_sort() to quiet uninitialized value warnings when running in
a Perl 7 environment.  Granted, the solution is not elegant, but it
improves the test output.

After applying this commit, we get this output in testing ppphtest.t:

    Failed 35/238 subtests

    Test Summary Report
    -------------------
    ../dist/Devel-PPPort/t/ppphtest.t (Wstat: 0 Tests: 209 Failed: 6)
      Failed tests:  7, 13, 62, 84, 163-164
      Parse errors: Bad plan.  You planned 238 tests but ran 209.

In p7 branch, we are also only running 209 tests.  I do not yet know why
we're running short of the count.  But the number of outright failures
is now much smaller: a reduction from 187 to 35.