recurse.t:  Kill prototypes on 2 subroutine declarations
This was necessary because with prototypes program was failing to
compile with errors like "Too many arguments for subroutine
MAX_DEPTH_HASH."

But note that 'perldiag' here speaks of subroutine signatures rather
than prototypes:  "A subroutine using a signature received more
arguments than permitted by the signature."  So this is an instance of
subroutine/prototype confusion.

At least now the program runs to completion -- but with failures.