t/comp/redef.t:  Skip all until prototype problems are resolved.
Add t/comp/redef-old.t to retain (as much as possible) of testing of Perl 5
functionality.

The purpose of t/comp/redef.t since the early days of Perl 5 has been to test
whether or not we get a specified warnings when a subroutine is redefined.  As
such, from the beginning it had to consider cases where prototyped subroutines
were redefined.  (It even uses an ok() testing subroutine, customized with
prototypes, written by Larry Wall in 1996!)  However, the program was never
updated to reflect the possiblity that sub prototypes could be declared with
the ':attribute()' syntax.  So while all its unit tests pass, they no longer
adequately cover the cases needed to be tested.

This complicates matters for Perl 7 semantics.  As of the date of this commit,
we are heavily dependent on the ':attribute()' syntax for avoiding confusion
with sub signatures, so that syntax needs to be tested.  On the other hand, if
we remove a 'use v5;' statement -- which we should, because the purpose of
this branch is to test Perl 7 semantics as much as possible -- then we very
quickly run into compile-time errors like this:

    Illegal character following sigil in a subroutine signature at comp/redef.t

So we're damned if we do and damned if we don't.

The solution offered by this commit is two-part:

1. Copy t/comp/redef.t from blead into this branch as t/comp/redef-old.t.
Modify it by adding 'use v5;' and the smallest possible number of additional
statements needed to get it to PASS in this branch.

2. Defer any attempt to get t/comp/redef.t in this branch to PASS until such
time as we sort out the confusion between subroutine prototypes and
signatures.  Fail the program immediately, clearly stating why we're doing so.

In this branch, while we certainly want all tests run clean, we don't want to
achieve that by papering over problems.