t/io/open.t: Adapt to honor strict-by-default
In Perl 5, this file had 'use warnings' turned on nearly throughout and
used that to capture and test the content of warnings in particular
blocks.  However, in Perl 5 this file only had 'use strict' turned on
for short stretches.

To be consistent with our development plan, in this commit we're testing
that the file compiles with strict-on-default, but we're restricting the
use of warnings to those blocks where we're actually concerned about the
text of the warnings.

For Objective 3 -- warnings-by-default -- we'll be more specifically
examine whether the program compiles and runs without warnings.

Some tidying of tabs into spaces.