t/test.pl:  Set @INC only as needed to be able to locate strict.pm
t/test.pl needs to run with relaxed strictures, at least for now.  To
run in an environment with strict-by-default, it needs to be able to say
"no strict;".  To say that, it needs to be able to locate 'strict.pm'
via @INC.  Hence, the "@INC = '../lib';" in the BEGIN block below.

PROBLEM:  Some test files that 'require' t/test.pl don't want that
assignment.

WORKAROUND:  Save the current state of @INC, hard-set it so that this
program can locate strict.pm.  Then restore the original @INC after
we've relaxed strictures.