t/porting/bench_selftest.t:  Improve setup of working directory
Most tests want to be run from the 't/' directory, so we're usually
looking to change into that directory from the top-level directory at
the start of program.

This test, however, is a wrapper around Porting/bench.pl.  So it wants
to start out from the top-level directory.  Make it so.