lib/dumpvar.pl:  needs to run 'no strict' for now
This progam is a classic example of mid-1990s Perl brilliance.  It is
the backbone of, among other things, the 'x' command in the Perl
debugger.  But it has global variables flying all over the place, so it
won't compile or run under 'use strict' in Perl 5, much less in a Perl
7-ish environment where strict is in every breath you take.