lib/DB-small.t:  Make test fail rather than be skipped
We've created this file as a place to hold a few particularly
problematic stanzas from lib/DB.t for diagnostic purposes.  It is not
intended as a permanent addition to the test suite.

Until we figure the problem out, we should expose the problem as a test
failure rather than hide it in a SKIP block.