t/lib/strict/*:  Update the data files for strict-by-default
In each test in a data file, in Perl 5 we're starting out with no strict
by default, so a 'use strict' statement means we're specifically testing
strictures.

Here, we start out with strict on by default, so to test the same
behavior as we did in a given test in Perl 5, we may have to inject a
'no strict' statement.

Where the EXPECT is expecting a line number in an error message, the
insertion/removal of a strict/no strict statement may necessitate an
increment/decrement in the line number expected.

Avoid use of $a and $b as they are special to 'sort' built-in function.

Correct one error in inline comment in common.pl