Update B::Concise tests to reflect strict being on by default in perl programs
Note that because -e tests do not behave this way, we need to also test it works
too. So we've added an option 'run_as_five' to test.pl so we can simulate that.

ext/B/t/optree_specials.t changed more than we expected and we don't totally
understand why. We will need further analysis to understand why.