win32/bin/search.pl:  Adapt to honor strict-by-default
This file is written in mid-1990s style perl, i.e., global variables all
over the place.  The amount of work needed to rewrite this with properly
scoped variables would exceed what is appropriate for a
proof-of-concept.  So for now we'll simply relax those strictures
obviously needed and be content to have the program compile.