cpan/ExtUtils-Install: sync in CPAN version 2.16
Then bring all files into compliance with strict-by-default.

The version of EU-I we previously had, 2.14, suffered from a race
condition.  Patches were submitted to PTG for this problem but not
released to CPAN (2.16) until it was too late to get them into
perl-5.32.0.  Hence, they were not part of Perl 7 at our initial branch
point, alpha-00.  In the last two days we got smoke-test failures due to
the race condition.  So to avoid misleading smoke test failure reports,
let's pull in 2.16 and make it strict-compliant.

Signed-off-by: James E Keenan <jkeenan@cpan.org>