cpan/Text-Balanced/t/09_gentag.t: Adapt to warnings-by-default
Per discussion in https://github.com/atoomic/perl/issues/301, we want to
avoid use of 'no warnings' in this module's tests as the upstream distro
on CPAN carries an implicit minimum Perl version of 5.003_096 (!).

Capturing the warning (without bothering to test for the content of the
capture) seems to be the least intrusive way of addressing this problem.