dist/Dumpvalue/t/Dumpvalue.t: Adapt to warnings-by-default
We were getting a "Useless use of a constant ("1") in void context"
warning.  Inspection showed that this was a legitimate warning, as
Test::More::is expects a scalar for both 'got' and 'expected', but the
test was coded such that the 'expected' was a list ( qw( 1 1 ) ).  So a
better test is to make a string out of the list returned by each of
'got' and 'expected'.