t/loc_tools.pl: Sync with Perl 5 blead
In part as a response to bug tickets we filed with Perl 5,
t/loc_tools.pl has had a lot of work done on it in recent months in
blead.  Let's pull this work in.

ext/POSIX/t/posix.t:  Apply correction requested in
https://github.com/Perl/perl5/issues/18245.

For: https://github.com/atoomic/perl/issues/305 (partial)